대댓글 달리는 위치 개선, 대댓글 있는 댓글 삭제시 '삭제된 메시지 입니다'라는 메시지로 변경

+++ src/main/java/study/hlf/service/CommentService.java
@@ -65,8 +65,23 @@ public class CommentService {
     @Transactional
     public boolean deleteComment(Long commentId, Long postId){
         try {
-            commentRepository.deleteById(commentId);
-            boardRepository.findById(postId).get().subCommentCount();
+            Optional<Comment> findComment = commentRepository.findById(commentId);
+            findComment.ifPresentOrElse((comment -> {
+                if(comment.getChildren().size() != 0){
+                    comment.changeStatus();
+                } else {
+                    Long parentId = comment.getParent().getId();
+                    if(comment.getParent().isDeleteStatus()
+                            && comment.getParent().getChildren().size() == 1){
+                        commentRepository.deleteById(commentId);
+                        commentRepository.deleteById(parentId);
+                        boardRepository.findById(postId).get().subCommentCount();
+                    } else {
+                        commentRepository.deleteById(commentId);
+                    }
+                    boardRepository.findById(postId).get().subCommentCount();
+                }
+            }), null);
         } catch (Exception e) {
             e.printStackTrace();
             return false;
